requirements.txt
requests==2.28.1
azure-storage-blob==12.10.0